net: openthread: rpc: Replace IPPROTO_RAW with ETH_P_ALL
The combination of AF_PACKET and IPPROTO_RAW is not correct as packet
sockets should only be used with IEEE 802.3 protocol numbers. As the
support for this combination will likely be soon removed from Zephyr,
switch to use DGRAM packet socket instead.

Using ETH_P_ALL is fine in this case, as OpenThread interface only
provides IPv6 packets anyway.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>